pixman
Struct Image

pub struct Image<'bits, 'alpha> { /* private fields */ }

Image holding some pixel data

Implementations

impl Image<'static, 'static>

pub fn new( format: FormatCode, width: usize, height: usize, clear: bool, ) -> Result<Self, CreateFailed>

Create a new image with the specified format and size

impl<'bits> Image<'bits, 'static>

pub fn from_slice_mut( format: FormatCode, width: usize, height: usize, bits: &'bits mut [u32], rowstride_bytes: usize, clear: bool, ) -> Result<Self, CreateFailed>

Create an image from some pre-allocated pixel data

pub unsafe fn from_raw_mut( format: FormatCode, width: usize, height: usize, bits: *mut u32, rowstride_bytes: usize, clear: bool, ) -> Result<Self, CreateFailed>

Create an image from some pre-allocated pixel data pointer

Safety

The caller is responsible to make sure the pointer stays valid for the lifetime of the created image.

impl<'bits, 'a> Image<'bits, 'a>

pub fn set_alpha_map<'alpha: 'a>( self, alpha_map: &'alpha Image<'_, 'static>, x: i16, y: i16, ) -> Image<'bits, 'alpha>

Set an alpha map that will be used when this image is used as a src in a blit operation

pub fn clear_alpha_map(self) -> Image<'bits, 'static>

Clear a previously set alpha map

impl Image<'_, '_>

pub fn width(&self) -> usize

Get the width of the image

pub fn height(&self) -> usize

Get the height of the image

pub fn stride(&self) -> usize

Get the stride of the image

pub fn depth(&self) -> usize

Get the depth of the image

pub fn format(&self) -> FormatCode

Get the format of the image

pub unsafe fn data(&self) -> *mut u32

Access the underlying pixel data

Safety

The pointer is valid for the lifetime of the image

pub fn fill_boxes( &mut self, op: Operation, color: impl Into<Color>, boxes: &[Box32], ) -> Result<(), OperationFailed>

Fill this image with the specified boxes and color

pub fn fill_rectangles( &mut self, op: Operation, color: impl Into<Color>, rects: &[Rectangle16], ) -> Result<(), OperationFailed>

Fill this image with the specified rectangles and color

pub fn composite( &mut self, operation: Operation, src: &ImageRef, mask: Option<&ImageRef>, src_loc: (i16, i16), mask_loc: (i16, i16), dest_loc: (i16, i16), size: (u16, u16), )

Composite the specified src image into this image

pub fn composite32( &mut self, operation: Operation, src: &ImageRef, mask: Option<&ImageRef>, src_loc: (i32, i32), mask_loc: (i32, i32), dest_loc: (i32, i32), size: (i32, i32), )

Composite the specified src image into this image

pub fn composite_triangles( &mut self, operation: Operation, src: &ImageRef, mask_format: FormatCode, src_loc: (i32, i32), dest_loc: (i32, i32), tris: &[Triangle], )

Composite the specified triangles into this image

pub fn composite_trapezoids( &mut self, operation: Operation, src: &ImageRef, mask_format: FormatCode, src_loc: (i32, i32), dest_loc: (i32, i32), traps: &[Trapezoid], )

Composite the specified trapezoids into this image

pub fn add_traps(&mut self, offset: (i16, i16), traps: &[Trap])

Add the specified traps to this image

pub fn add_trapezoids(&mut self, offset: (i16, i32), traps: &[Trapezoid])

Add the specified trapezoids to this image

pub fn add_triangles(&mut self, offset: (i32, i32), tris: &[Triangle])

Add the specified triangles to this image

pub fn compute_composite_region( &self, src: &ImageRef, mask: Option<&ImageRef>, src_loc: (i16, i16), mask_loc: (i16, i16), dest_loc: (i16, i16), size: (u16, u16), ) -> Option<Region16>

Compute the composite region for the specified src

pub fn rasterize_edges( &mut self, l: Edge, r: Edge, t: impl Into<Fixed>, b: impl Into<Fixed>, )

Rasterize the specified edges

pub fn rasterize_trapezoid(&mut self, trap: Trapezoid, offset: (i32, i32))

Rasterize the specified trapezoids

impl Image<'_, '_>

pub unsafe fn from_ptr(ptr: *mut pixman_image_t) -> Self

Initialize the image from a raw pointer

Safety

The pointer is expected to be valid and have a ref-count of at least one. Ownership of the pointer is transferred and unref will be called on drop.

Any other references to the pixman_image_t must not be mutated while this Image exists, including changes to the reference count.