Crate pixman

From the official pixman docs:

Pixman is a low-level software library for pixel manipulation, providing features such as image compositing and trapezoid rasterization. Important users of pixman are the cairo graphics library and the X server.

Pixman is implemented as a library in the C programming language. It runs on many platforms, including Linux, BSD Derivatives, MacOS X, and Windows.

Pixman is free and open source software. It is available to be redistributed and/or modified under the terms of the MIT license.

Re-exports

pub use pixman_sys as ffi;

Structs

Color

Rgba color in the range of u16::MIN to u16::MAX

ConicalGradient

Conical gradient image

CreateFailed

Resource creation failed

Edge

Defines a single edge

FTransform

Floating-point transform

FVector

Floating-point vector

Fixed

Fixed-point value

GradientStop

Gradient-stop

Image

Image holding some pixel data

ImageRef

A reference to a raw image

Line

Single line

LinearGradient

Linear gradient image

OperationFailed

The requested operation failed

Point

A single point

RadialGradient

Radial gradient image

Region16

16bit region

Region32

32bit region

Solid

Solid color image

Span

A single span

Transform

Fixed-point transform

TransformConvertError

Failed to init Transform from FTransform

Trap

A single trap

Trapezoid

A trapezoid

Triangle

A triangle

UnknownDither

The dither operation is unknown

UnknownFilter

The filter operation is unknown

UnknownOperation

The operation is unknown

UnknownRepeat

The repeat operations is unknown

UnsupportedDrmFourcc

The provided drm-fourcc has no matching format code

UnsupportedFormatCode

The provided format code has no matching drm-fourcc

Vector

A single vector

Enums

Dither

Defiens the possible dither operations

Filter

Defines the possible filter operations

FormatCode

Possible format codes

Operation

Defines the possible operations for composition

Overlap

Describes overlap of a region with a rectangle

Repeat

Defines the possible repeat operations

Functions

blit

Blit the src into the dst with the specified values

fill

Fill the provided bits with the provided values

sample_ceil_y

Compute the smallest value greater than or equal to y which is on a grid row.

sample_floor_y

Compute the largest value strictly less than y which is on a grid row.

Type Aliases

Box16

16bit Box

Box32

32bit Box

Rectangle16

16bit Rectangle

Rectangle32

32bit Rectangle